+++ packages/core-js/stage/1.js
@@ -1,3 +1,4 @@
+require('../proposals/array-filtering');
 require('../proposals/array-last');
 require('../proposals/collection-methods');
 require('../proposals/collection-of-from');