+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Changelog
 ##### Unreleased
-- Added [`{ Number, BigInt }.range` methods stage 1 proposal](https://github.com/tc39/proposal-Number.range)
+- Added [`Number.range` stage 1 proposal](https://github.com/tc39/proposal-Number.range)
+  - `Number.range`
+  - `BigInt.range`
+- Added [array filtering stage 1 proposal](https://github.com/tc39/proposal-array-filtering)
+  - `Array#filterOut`
+  - `%TypedArray%#filterOut`
 - Added the missed `(es|stable)/instance/replace-all` entries
 - Updated compat data mapping for Opera - from Opera 69, the difference with Chrome versions increased to 14
 - Compat data mapping for modern Android WebView to Chrome moved from targets parser directly to compat data